Weather initial BackEnd Setup

+++ AirQualityMonitor/core/src/main/java/com/example/appCore/domain/PeakFlow.kt
@@ -1,5 +1,5 @@
 package com.example.appCore.domain
 
 //Reading (260 - 270 units, best is 300 units)
-class PeakFlow(val peakFlow: Float) {
+class PeakFlow(val peakFlow: Int) {
 }
